requirements.txt
pysui
python-dotenv==1.0.1
openai
